Добавил скрипты и Dockerfile'ы для компиляции и запуска компилятора в Docker

+++ builder/setup-builder.sh
@@ -25,14 +25,31 @@ install_clang() {
     wget -O - https://apt.llvm.org/llvm-snapshot.gpg.key | apt-key add -
     add_apt_repository llvm-toolchain-stretch http://apt.llvm.org/stretch/ llvm-toolchain-stretch main
     apt-get -qq update
-    apt-get -qq install -y libstdc++-6-dev clang-3.9 llvm-3.9-dev cmake
+    apt-get -qq install -y libstdc++-6-dev clang-3.9 llvm-3.9-dev
 
-    ln -s /usr/bin/clang-3.9 /usr/bin/cc
+    if ! [ -f /usr/bin/cc ]; then
+        ln -s /usr/bin/clang-3.9 /usr/bin/cc
+    fi
     ln -s /usr/bin/clang-3.9 /usr/bin/clang
+    if ! [ -f /usr/bin/c++ ]; then
+        ln -s /usr/bin/clang++-3.9 /usr/bin/c++
+    fi
     ln -s /usr/bin/clang++-3.9 /usr/bin/clang++
-    ln -s /usr/bin/clang++-3.9 /usr/bin/c++
-    ln -s /usr/bin/clang++-3.9 /usr/bin/c++
+}
+
+install_cmake() {
+    echo 3. install cmake
+    apt-get -qq update
+    apt-get -qq install -y  cmake
+}
+
+install_dev_libraries() {
+    echo 4. install boost
+    apt-get -qq update
+    apt-get -qq install -y  libboost-all-dev zlib1g-dev
 }
 
 prepare_install
 install_clang
+install_cmake
+install_dev_libraries
